Update JOINer
Added a section to check if the table name already exists. Previously the proc always created or replaced the table. But BQ can't create a table in a multi-statement transaction block. This allows users to pre-create the table, in which case the proc will use an INSERT statement instead and can then be used in a transaction block.