Update UPSERTer
Added a session name and included it as a column in the temp table. This is needed because, as of now, when you run a CREATE OR REPLACE statement for a temp table multiple times within the same transaction block, it isn't actually replacing the temp table in the second (or greater) run. This means that in cases where this proc is run multiple times within the same transaction block, it would error out because there were too many rows in the temp table. This appears to be a bug in BQ.