Update JOINer
Created two new versions of joiner. They account for situations where the same source record has multiple database transactions in between batches. This new version only retrieves the most recent version for the right side tables (this is not done on the left, or main, table). There is a version that includes the operation flag as well in order to filter out deletes.